Move filterCourses function from util to selector
