+++ src/selectors/course.js
@@ -1,5 +1,5 @@
 import {createSelector} from 'reselect';
-import {filterCourses} from '../util';
+import {filterLessons} from '../util';
 
 const getAllCourses = (state) => state.allCourses;
 const getFilter = (state) => state.filter;
@@ -7,6 +7,16 @@ const getFilter = (state) => state.filter;
 export const getFilteredCourses = createSelector(
   [getFilter, getAllCourses],
   (filter={}, allCourses=[]) => {
-    return filterCourses(allCourses, filter);
+    const coursesWithFilteredLessons = allCourses.map(course => {
+      const newCourse = {...course};
+      newCourse.lessons = filterLessons(course.lessons, filter);
+      return newCourse;
+    });
+
+    // Find courses that have at least one lesson that matches filter
+    return coursesWithFilteredLessons.filter((course) => {
+      return course.lessons.length > 0;
+    });
+    // return filterCourses(allCourses, filter);
   }
 );
