+++ src/util.js
@@ -56,26 +56,6 @@ export function getCourses(lessonContext, iconContext) {
 }
 
 /**
- * Get courses that has at least one lesson that matches filter
- * @param {Array} courses
- * @param {Object} filter
- * @returns {Array} filtered courses
- */
-export function filterCourses(courses, filter) {
-  const coursesWithFilteredLessons = courses.map(course => {
-    const newCourse = {...course};
-    newCourse.lessons = filterLessons(course.lessons, filter);
-    return newCourse;
-  });
-
-  // Find courses that have at least one lesson that matches filter
-  return coursesWithFilteredLessons.filter((course) => {
-    return course.lessons.length > 0;
-  });
-
-}
-
-/**
  * Get lessons that have all tags that exist in filter
  * @param {Array} lessons
  * @param {Object} filter
